+++ flowauth/backend/tests/test_token_generation.py
@@ -4,8 +4,10 @@
 
 import datetime
 
-import jwt
 import pytest
+import jwt
+from flowauth.token_management import generate_token as flowauth_generate_token
+from flowkit_jwt_generator import generate_token as jwt_generator_generate_token
 
 from pytest import approx
 
@@ -59,10 +61,10 @@ def test_token_generation(client, auth, app):
     assert 200 == response.status_code
     token_json = response.get_json()
     decoded_token = jwt.decode(
-        audience="DUMMY_SERVER_A_KEY",
         jwt=token_json["token"].encode(),
         key=app.config["PUBLIC_JWT_SIGNING_KEY"],
         algorithms=["RS256"],
+        audience="DUMMY_SERVER_A",
     )
     assert {
         "DUMMY_ROUTE_A": {
@@ -125,3 +127,43 @@ def test_token_rejected_for_bad_right_claim(client, auth, app):
         b"You do not have access to DUMMY_ROUTE_A:poll on DUMMY_SERVER_A"
         in response.get_data()
     )
+
+
+def test_against_general_generator():
+    """Test that the token generator in FlowAuth and the one in flowkit-jwt-generator produce same results."""
+    flowauth_token = jwt.decode(
+        flowauth_generate_token(
+            username="TEST_USER",
+            secret="SECRET",
+            lifetime=datetime.timedelta(5),
+            flowapi_identifier="TEST_SERVER",
+            claims={
+                "daily_location": {
+                    "permissions": {"run": True},
+                    "spatial_aggregation": ["admin3"],
+                }
+            },
+        ),
+        key="SECRET",
+        audience="TEST_SERVER",
+    )
+    generator_token = jwt.decode(
+        jwt_generator_generate_token(
+            username="TEST_USER",
+            secret="SECRET",
+            lifetime=datetime.timedelta(5),
+            flowapi_identifier="TEST_SERVER",
+            claims={
+                "daily_location": {
+                    "permissions": {"run": True},
+                    "spatial_aggregation": ["admin3"],
+                }
+            },
+        ),
+        key="SECRET",
+        audience="TEST_SERVER",
+    )
+    assert generator_token["aud"] == flowauth_token["aud"]
+    assert generator_token["user_claims"] == flowauth_token["user_claims"]
+    assert generator_token["identity"] == flowauth_token["identity"]
+    assert generator_token.keys() == flowauth_token.keys()