+++ flowauth/frontend/src/ServerAdminDetails.js
@@ -15,6 +15,7 @@ import InputAdornment from "@material-ui/core/InputAdornment";
 import TextField from "@material-ui/core/TextField";
 import SubmitButtons from "./SubmitButtons";
 import ServerAggregationUnits from "./ServerAggregationUnits";
+import ErrorDialog from "./ErrorDialog";
 import { generate } from "generate-password";
 import {
   getAllAggregationUnits,
@@ -33,7 +34,12 @@ class ServerAdminDetails extends React.Component {
     rights: {},
     max_life: 1440,
     latest_expiry: new Date(new Date().getTime() + 24 * 60 * 60 * 1000),
-    edit_mode: false
+    edit_mode: false,
+    name_helper_text: "",
+    key_helper_text: "",
+    maxlife_helper_text: "",
+    pageError: false,
+    errors: { message: "" }
   };
 
   handleSubmit = () => {
@@ -43,33 +49,42 @@ class ServerAdminDetails extends React.Component {
       latest_expiry,
       max_life,
       rights,
+      name_helper_text,
+      key_helper_text,
+      maxlife_helper_text
     } = this.state;
     const { item_id, onClick } = this.props;
-    var task;
-    if (edit_mode) {
-      task = editServer(
-        item_id,
-        name,
-        new Date(latest_expiry).toISOString(),
-        max_life
-      );
-    } else {
-      task = createServer(
-        name,
-        new Date(latest_expiry).toISOString(),
-        max_life
-      );
+    if (name && name_helper_text === "" && key_helper_text === "" && max_life && maxlife_helper_text === "") {
+      var task;
+      if (edit_mode) {
+        task = editServer(
+          item_id,
+          name,
+          new Date(latest_expiry).toISOString(),
+          max_life
+        );
+      } else {
+        task = createServer(
+          name,
+          new Date(latest_expiry).toISOString(),
+          max_life
+        );
+      }
+      task
+        .then(json => {
+          return editServerCapabilities(json.id, rights);
+        })
+        .then(json => {
+          onClick();
+        })
+        .catch(err => {
+          if (err.code === 400) {
+            this.setState({ pageError: true, errors: err });
+          } else {
+            this.setState({ hasError: true, error: err })
+          }
+        });
     }
-    task
-      .then(json => {
-        return editServerCapabilities(json.id, rights);
-      })
-      .then(json => {
-        onClick();
-      })
-      .catch(err => {
-        this.setState({ hasError: true, error: err });
-      });
   };
 
   fieldHasError = field => {
@@ -79,6 +94,7 @@ class ServerAdminDetails extends React.Component {
       return false;
     }
   };
+
 
 
   handleDateChange = date => {
@@ -105,13 +121,53 @@ class ServerAdminDetails extends React.Component {
 
   handleTextChange = name => event => {
     this.setState({
-      [name]: event.target.value
+      pageError: false,
+      errors: ""
     });
-  };
+    var state = {
+      [name]: event.target.value
+    };
+    if (name === "name") {
+      var letters = /^[A-Za-z0-9_]+$/;
+      let servername = event.target.value;
+      if (servername.match(letters) && servername.length <= 120) {
+        state = Object.assign(state, {
+          name_helper_text: ""
+        });
+      } else if (servername.length == 0) {
+        state = Object.assign(state, {
+          name_helper_text: "Server name can not be blank."
+        });
+      } else if (!servername.match(letters)) {
+        state = Object.assign(state, {
+          name_helper_text: "Server name may only contain letters, numbers and underscores."
+        });
+      } else {
+        state = Object.assign(state, {
+          name_helper_text: "Server name must be 120 characters or less."
+        });
+      }
+    }
+
+
+    if (name === "max_life") {
+      let maxlife = event.target.value;
+      if (maxlife.length == 0) {
+        state = Object.assign(state, {
+          maxlife_helper_text: "Maximum lifetime minutes can not be blank."
+        });
+      } else {
+        state = Object.assign(state, {
+          maxlife_helper_text: ""
+        });
+      }
+    }
+    this.setState(state);
+  }
 
   componentDidMount() {
     const { item_id } = this.props;
-    var name, rights, secret_key;
+    var name, rights;
     if (item_id && true) {
       getServer(item_id)
         .then(json => {
@@ -163,13 +219,17 @@ class ServerAdminDetails extends React.Component {
         </Grid>
         <Grid item xs={6}>
           <TextField
-            error={this.fieldHasError("name")}
-            id="standard-name"
+            // error={this.fieldHasError("name")}
+            placeholder="Maximum 120 characters"
+            id="name"
             label="Name"
             className={classes.textField}
             value={name}
             onChange={this.handleTextChange("name")}
             margin="normal"
+            required={true}
+            error={this.state.name_helper_text}
+            helperText={(this.state.name_helper_text)}
           />
         </Grid>
         <Grid item xs={6} />
@@ -193,13 +253,16 @@ class ServerAdminDetails extends React.Component {
         </Grid>
         <Grid item xs={6}>
           <TextField
-            id="standard-name"
+            id="max-life"
             label="Maximum lifetime (minutes)"
             className={classes.textField}
             type="number"
             value={max_life}
             onChange={this.handleTextChange("max_life")}
             margin="normal"
+            required={true}
+            error={this.state.maxlife_helper_text}
+            helperText={this.state.maxlife_helper_text}
           />
         </Grid>
         <Grid item xs={12}>
@@ -236,6 +299,10 @@ class ServerAdminDetails extends React.Component {
             }}
           />
         ))}
+        <ErrorDialog
+          open={this.state.pageError}
+          message={this.state.errors.message}
+        />
         <SubmitButtons handleSubmit={this.handleSubmit} onClick={onClick} />
       </React.Fragment>
     );