+++ flowkit_jwt_generator/flowkit_jwt_generator/__init__.py
@@ -8,12 +8,6 @@ from .jwt import (
     aggregation_types,
     permissions_types,
     generate_keypair,
+    load_private_key,
+    load_public_key,
 )
-
-__all__ = [
-    "generate_token",
-    "generate_keypair",
-    "get_all_claims_from_flowapi",
-    "aggregation_types",
-    "permissions_types",
-]