+++ flowmachine/flowmachine/core/query_info_lookup.py
@@ -2,7 +2,7 @@ import rapidjson
 import structlog
 from redis import StrictRedis
 
-logger = structlog.get_logger(__name__)
+logger = structlog.get_logger("flowmachine.debug", submodule=__name__)
 
 
 class QueryInfoLookupError(Exception):